Remove unused BEAT engine code
