--- src/blab/solvers/afterburner_backend.py
@@ -1,25 +1,0 @@
-"""Compatibility shim for the renamed BEAT Engine backend."""
-
-from blab.solvers.beat_engine_backend import (
-    DEFAULT_AFTERBURNER_PROJECT,
-    DEFAULT_AFTERBURNER_SOLVER_SCRIPT,
-    DEFAULT_BEAT_ENGINE_PROJECT,
-    DEFAULT_BEAT_ENGINE_SOLVER_SCRIPT,
-    DEFAULT_JULIA_PROJECT,
-    DEFAULT_JULIA_SOLVER_SCRIPT,
-    AfterburnerBackend,
-    AfterburnerSession,
-    AfterburnerWorkerProcess,
-    BeatEngineBackend,
-    BeatEngineCpuBackend,
-    BeatEngineCudaBackend,
-    BeatEngineSession,
-    BeatEngineWorkerProcess,
-    JuliaLocalBackend,
-    JuliaLocalSession,
-    JuliaWorkerProcess,
-    _resolve_julia_threads,
-    shutdown_afterburner_workers,
-    shutdown_beat_engine_workers,
-    shutdown_julia_workers,
-)
