+++ src/blab/solvers/julia_local/src/BeatEngineCudaOperators.jl
@@ -51,115 +51,6 @@ _cuda_timed_stage!(thunk, timing, name::String) = _cuda_timed_stage!(timing, nam
 
 _regular_quadrature_threads(rule_count::Int) = 16
 
-function _cuda_regular_real_buffers(::Type{T}, p1_dof_count::Int, dp0_dof_count::Int) where {T<:AbstractFloat}
-    return (
-        slp_re=CUDA.zeros(T, p1_dof_count, dp0_dof_count),
-        slp_im=CUDA.zeros(T, p1_dof_count, dp0_dof_count),
-        adj_re=CUDA.zeros(T, p1_dof_count, dp0_dof_count),
-        adj_im=CUDA.zeros(T, p1_dof_count, dp0_dof_count),
-        dlp_re=CUDA.zeros(T, p1_dof_count, p1_dof_count),
-        dlp_im=CUDA.zeros(T, p1_dof_count, p1_dof_count),
-        hyp_re=CUDA.zeros(T, p1_dof_count, p1_dof_count),
-        hyp_im=CUDA.zeros(T, p1_dof_count, p1_dof_count),
-    )
-end
-
-function _cuda_fill_regular_real_buffers!(buffers, value)
-    fill!(buffers.slp_re, value)
-    fill!(buffers.slp_im, value)
-    fill!(buffers.adj_re, value)
-    fill!(buffers.adj_im, value)
-    fill!(buffers.dlp_re, value)
-    fill!(buffers.dlp_im, value)
-    fill!(buffers.hyp_re, value)
-    fill!(buffers.hyp_im, value)
-    CUDA.synchronize()
-    return nothing
-end
-
-function _cuda_free_regular_real_buffers!(buffers)
-    CUDA.unsafe_free!(buffers.slp_re)
-    CUDA.unsafe_free!(buffers.slp_im)
-    CUDA.unsafe_free!(buffers.adj_re)
-    CUDA.unsafe_free!(buffers.adj_im)
-    CUDA.unsafe_free!(buffers.dlp_re)
-    CUDA.unsafe_free!(buffers.dlp_im)
-    CUDA.unsafe_free!(buffers.hyp_re)
-    CUDA.unsafe_free!(buffers.hyp_im)
-    return nothing
-end
-
-function _launch_regular_split_balanced_atomic_kernel!(
-    slp_re,
-    slp_im,
-    dlp_re,
-    dlp_im,
-    adj_re,
-    adj_im,
-    hyp_re,
-    hyp_im,
-    d_face_vertices,
-    d_normals,
-    d_areas,
-    d_faces,
-    d_curls,
-    d_test_indices,
-    d_trial_indices,
-    d_rule_points,
-    d_rule_weights,
-    k::T,
-    p1_dof_count::Int,
-    face_count::Int,
-    rule_count::Int,
-    total_pairs::Int,
-    threads::Int,
-) where {T<:AbstractFloat}
-    shmem = threads * 24 * sizeof(T)
-    CUDA.@cuda threads=threads blocks=total_pairs shmem=shmem _cuda_regular_quadrature_slp_hyp_kernel!(
-        slp_re,
-        slp_im,
-        hyp_re,
-        hyp_im,
-        d_face_vertices,
-        d_normals,
-        d_areas,
-        d_faces,
-        d_curls,
-        d_test_indices,
-        d_trial_indices,
-        d_rule_points,
-        d_rule_weights,
-        k,
-        p1_dof_count,
-        face_count,
-        rule_count,
-        total_pairs,
-        1,
-    )
-
-    CUDA.@cuda threads=threads blocks=total_pairs shmem=shmem _cuda_regular_quadrature_dlp_adjoint_kernel!(
-        dlp_re,
-        dlp_im,
-        adj_re,
-        adj_im,
-        d_face_vertices,
-        d_normals,
-        d_areas,
-        d_faces,
-        d_test_indices,
-        d_trial_indices,
-        d_rule_points,
-        d_rule_weights,
-        k,
-        p1_dof_count,
-        face_count,
-        rule_count,
-        total_pairs,
-        1,
-    )
-    return nothing
-end
-
 function _launch_regular_split_balanced_multipair_atomic_kernel!(
     slp_re,
     slp_im,
@@ -230,93 +121,6 @@ function _launch_regular_split_balanced_multipair_atomic_kernel!(
         rule_count,
         total_pairs,
         pairs_per_block,
-    )
-    return nothing
-end
-
-function _launch_regular_split_slp_hyp_separate_atomic_kernel!(
-    slp_re,
-    slp_im,
-    dlp_re,
-    dlp_im,
-    adj_re,
-    adj_im,
-    hyp_re,
-    hyp_im,
-    d_face_vertices,
-    d_normals,
-    d_areas,
-    d_faces,
-    d_curls,
-    d_test_indices,
-    d_trial_indices,
-    d_rule_points,
-    d_rule_weights,
-    k::T,
-    p1_dof_count::Int,
-    face_count::Int,
-    rule_count::Int,
-    total_pairs::Int,
-    threads::Int,
-) where {T<:AbstractFloat}
-    slp_shmem = threads * 6 * sizeof(T)
-    CUDA.@cuda threads=threads blocks=total_pairs shmem=slp_shmem _cuda_regular_quadrature_slp_kernel!(
-        slp_re,
-        slp_im,
-        d_face_vertices,
-        d_areas,
-        d_faces,
-        d_test_indices,
-        d_trial_indices,
-        d_rule_points,
-        d_rule_weights,
-        k,
-        p1_dof_count,
-        face_count,
-        rule_count,
-        total_pairs,
-    )
-
-    hyp_shmem = threads * 18 * sizeof(T)
-    CUDA.@cuda threads=threads blocks=total_pairs shmem=hyp_shmem _cuda_regular_quadrature_hyp_kernel!(
-        hyp_re,
-        hyp_im,
-        d_face_vertices,
-        d_normals,
-        d_areas,
-        d_faces,
-        d_curls,
-        d_test_indices,
-        d_trial_indices,
-        d_rule_points,
-        d_rule_weights,
-        k,
-        p1_dof_count,
-        face_count,
-        rule_count,
-        total_pairs,
-    )
-
-    dlp_adjoint_shmem = threads * 24 * sizeof(T)
-    CUDA.@cuda threads=threads blocks=total_pairs shmem=dlp_adjoint_shmem _cuda_regular_quadrature_dlp_adjoint_kernel!(
-        dlp_re,
-        dlp_im,
-        adj_re,
-        adj_im,
-        d_face_vertices,
-        d_normals,
-        d_areas,
-        d_faces,
-        d_test_indices,
-        d_trial_indices,
-        d_rule_points,
-        d_rule_weights,
-        k,
-        p1_dof_count,
-        face_count,
-        rule_count,
-        total_pairs,
-        1,
     )
     return nothing
 end
