+++ src/blab/solvers/julia_local_backend.py
@@ -1,15 +1,10 @@
 """Compatibility imports for the BEAT Engine backend."""
 
 from blab.solvers.beat_engine_backend import (
-    DEFAULT_AFTERBURNER_PROJECT,
-    DEFAULT_AFTERBURNER_SOLVER_SCRIPT,
     DEFAULT_BEAT_ENGINE_PROJECT,
     DEFAULT_BEAT_ENGINE_SOLVER_SCRIPT,
     DEFAULT_JULIA_PROJECT,
     DEFAULT_JULIA_SOLVER_SCRIPT,
-    AfterburnerBackend,
-    AfterburnerSession,
-    AfterburnerWorkerProcess,
     BeatEngineBackend,
     BeatEngineCpuBackend,
     BeatEngineCudaBackend,
@@ -19,7 +14,6 @@ from blab.solvers.beat_engine_backend import (
     JuliaLocalSession,
     JuliaWorkerProcess,
     _resolve_julia_threads,
-    shutdown_afterburner_workers,
     shutdown_beat_engine_workers,
     shutdown_julia_workers,
 )
